fix data types for timeoutread(), timeoutwrite()

+++ base/timeoutread.h
@@ -1,5 +1,8 @@
 /*
  * $Log: timeoutread.h,v $
+ * Revision 1.4  2021-03-03 23:56:32+05:30  Cprogrammer
+ * fix data types
+ *
  * Revision 1.3  2004-10-11 14:15:25+05:30  Cprogrammer
  * added function prototypes
  *
@@ -9,7 +12,8 @@
  */
 #ifndef TIMEOUTREAD_H
 #define TIMEOUTREAD_H
+#include <sys/types.h>
 
-int             timeoutread(int, int, char *, int);
+ssize_t         timeoutread(long, int, char *, size_t);
 
 #endif
